requirements.txt
openai==0.28
langchain
pypdf
python-dotenv
faiss-cpu==1.9.0.post1
langchain-community
tiktoken>=0.4.0






